Fix compilation warnings.

+++ src/aSSA/Collectives.cpp
@@ -74,7 +74,7 @@ bool isCollective(const llvm::Function *F) {
 }
 
 bool isMpiCollective(int color) {
-  if (color >= v_coll.size())
+  if (color >= (int) v_coll.size())
     return false;
 
   for (unsigned i=0; i<MPI_v_coll.size(); ++i) {
@@ -86,7 +86,7 @@ bool isMpiCollective(int color) {
 }
 
 bool isOmpCollective(int color) {
-  if (color >= v_coll.size())
+  if (color >= (int) v_coll.size())
     return false;
 
   for (unsigned i=0; i<OMP_v_coll.size(); ++i) {
@@ -98,7 +98,7 @@ bool isOmpCollective(int color) {
 }
 
 bool isUpcCollective(int color) {
-  if (color >= v_coll.size())
+  if (color >= (int) v_coll.size())
     return false;
 
   for (unsigned i=0; i<UPC_v_coll.size(); ++i) {
@@ -110,7 +110,7 @@ bool isUpcCollective(int color) {
 }
 
 bool isCudaCollective(int color) {
-  if (color >= v_coll.size())
+  if (color >= (int) v_coll.size())
     return false;
 
   for (unsigned i=0; i<CUDA_v_coll.size(); ++i) {
